Move venue booking components to venues folder
